+++ R/as_data_frame.R
@@ -44,19 +44,19 @@
 #' @param x \[`dynamitefit`]\cr The model fit object.
 #' @param row.names Ignored.
 #' @param optional Ignored.
+#' @param types \[`character()`]\cr Type(s) of the parameters for which the
+#'   samples should be extracted. See details of possible values. Default is
+#'   all values listed in details except spline coefficients `omega`.
+#'   This argument is mutually exclusive with `parameters`.
 #' @param parameters \[`character()`]\cr Parameter(s) for which the samples
 #'   should be extracted. Possible options can be found with function
-#'   `get_parameter_names()`. Default is all parameters of specific type for all
-#'   responses.
+#'   `get_parameter_names()`. Default is all parameters of specific type for
+#'   all responses. This argument is mutually exclusive with `types`.
 #' @param responses \[`character()`]\cr Response(s) for which the samples
 #'   should be extracted. Possible options are elements of
 #'   `unique(x$priors$response)`, and the default is this entire vector.
 #'    Ignored if the argument `parameters` is supplied.
-#' @param types \[`character()`]\cr Type(s) of the parameters for which the
-#'   samples should be extracted. See details of possible values. Default is
-#'   all values listed in details except spline coefficients `omega`,
 #'   `omega_alpha`, and `omega_psi`. See also [dynamite::get_parameter_types()].
-#'    Ignored if the argument `parameters` is supplied.
 #' @param times \[`double()`]\cr Time point(s) to keep. If `NULL`
 #'   (the default), all time points are kept.
 #' @param groups \[`character()`] Group name(s) to keep. If `NULL`
@@ -107,23 +107,24 @@
 #' )
 #'
 as.data.frame.dynamitefit <- function(x, row.names = NULL, optional = FALSE,
-                                      parameters = NULL, responses = NULL,
-                                      types = NULL, times = NULL, groups = NULL,
+                                      types = NULL, parameters = NULL,
+                                      responses = NULL,
+                                      times = NULL, groups = NULL,
                                       summary = FALSE, probs = c(0.05, 0.95),
                                       include_fixed = TRUE, ...) {
   out <- as.data.table.dynamitefit(
-    x,
+    x = x,
     keep.rownames = FALSE,
-    row.names,
-    optional,
-    parameters,
-    responses,
-    types,
-    times,
-    groups,
-    summary,
-    probs,
-    include_fixed,
+    row.names = row.names,
+    optional = optional,
+    parameters = parameters,
+    responses = responses,
+    types = types,
+    times = times,
+    groups = groups,
+    summary = summary,
+    probs = probs,
+    include_fixed = include_fixed,
     ...
   )
   tibble::tibble(data.table::setDF(out))