+++ R/coef.R
@@ -5,17 +5,6 @@
 #' @export
 #' @family output
 #' @param object \[`dynamitefit`]\cr The model fit object.
-#' @param type  \[`character(1)`]\cr Either `beta` (the default) for
-#'   time-invariant coefficients, `delta` for time-varying coefficients,
-#'   `nu` for random effects, `lambda` for factor loadings, or `psi` for
-#'   latent factor. Ignored if the argument `parameters` is supplied.
-#' @param include_alpha \[`logical(1)`]\cr If `TRUE` (default), extracts also
-#'   time-invariant intercept term alpha if time-invariant parameters beta are
-#'   extracted, and time-varying alpha if time-varying delta are extracted.
-#'   Ignored if the argument `parameters` is supplied.
-#' @param include_cutpoints \[`logical(1)`]\cr If `TRUE` (default), plots also
-#'   the cutpoints if such parameters exists in the model (either time-varying
-#'   or time-invariant).
 #' @param summary \[`logical(1)`]\cr If `TRUE` (default), returns posterior
 #'   mean, standard deviation, and posterior quantiles (as defined by the
 #'   `probs` argument) for all parameters. If `FALSE`, returns the
@@ -31,12 +20,9 @@
 #' deltas <- coef(gaussian_example_fit, type = "delta")
 #'
 coef.dynamitefit <- function(object,
-                             parameters = NULL,
-                             type = c("beta", "delta", "nu", "lambda", "psi"),
+                             types = NULL, parameters = NULL,
                              responses = NULL, times = NULL, groups = NULL,
-                             summary = TRUE, probs = c(0.05, 0.95),
-                             include_alpha = TRUE,
-                             include_cutpoints = TRUE, ...) {
+                             summary = TRUE, probs = c(0.05, 0.95), ...) {
   stopifnot_(
     !missing(object),
     "Argument {.arg object} is missing."
@@ -46,58 +32,24 @@ coef.dynamitefit <- function(object,
     "Argument {.arg object} must be a {.cls dynamitefit} object."
   )
   if (is.null(parameters)) {
-    type <- onlyif(is.character(type), tolower(type))
-    type <- try(
-      match.arg(type, c("beta", "delta", "nu", "lambda", "psi")),
-      silent = TRUE
-    )
+    types <- onlyif(is.character(types), tolower(types))
+    invalid_types <- types[!types %in% all_types]
     stopifnot_(
-      !inherits(type, "try-error"),
-      "Argument {.arg type} must be either {.val beta}, {.val delta},
-      {.val nu}, {.val lambda}, or {.val psi}."
-    )
-    stopifnot_(
-      checkmate::test_flag(x = include_alpha),
-      "Argument {.arg include_alpha} must be single {.cls logical} value."
-    )
-    stopifnot_(
-      checkmate::test_flag(x = include_cutpoints),
-      "Argument {.arg include_cutpoints} must be a single {.cls logical} value."
+      length(invalid_types) == 0L,
+      c(
+        "Argument {.arg types} must contain only valid types.",
+        `x` = "Invalid types found: {invalid_types}"
+      )
     )
   }
-  if (is.null(parameters) && type %in% c("beta", "delta")) {
-    types <- c(
-      type,
-      onlyif(include_alpha, "alpha"),
-      onlyif(include_cutpoints, "cutpoints")
-    )
-    out <- as.data.frame.dynamitefit(
-      object,
-      parameters = parameters,
-      responses = responses,
-      types = types,
-      times = times,
-      groups = groups,
-      summary = summary,
-      probs = probs
-    )
-    # remove extra alphas
-    if (identical(type, "delta")) {
-      out <- out[!is.na(out$time), ]
-    } else {
-      out <- out[is.na(out$time), ]
-    }
-  } else {
-    out <- as.data.frame.dynamitefit(
-      object,
-      parameters = parameters,
-      responses = responses,
-      types = type,
-      times = times,
-      groups = groups,
-      summary = summary,
-      probs = probs
-    )
-  }
-  out
+  as.data.frame.dynamitefit(
+    x = object,
+    types = types,
+    parameters = parameters,
+    responses = responses,
+    times = times,
+    groups = groups,
+    summary = summary,
+    probs = probs
+  )
 }